Переделана проверка формы перед отправкой
