+++ css/style.css
@@ -394,6 +394,11 @@ fieldset {
           box-shadow: 0 0 2px 2px #ff6547;
 }
 
+.ad-form__element input:invalid,
+.ad-form__element select:invalid {
+  border-color: red;
+}
+
 .ad-form__element--time {
   -ms-flex-preferred-size: 50%;
       flex-basis: 50%;
