+++ js/form.js
@@ -32,70 +32,3 @@ const makePageActive = () => {
 
 makePageInactive();
 makePageActive();
-
-//Подсказка на короткое и длинное значение.
-const MIN_TITLE_LENGTH = 30;
-const MAX_TITLE_LENGTH = 100;
-
-const offerTitleInput = document.querySelector('#title');
-offerTitleInput.addEventListener('input', () => {
-  const titleLength = offerTitleInput.value.length;
-
-  if (titleLength < MIN_TITLE_LENGTH) {
-    offerTitleInput.setCustomValidity(`Ещё ${MIN_TITLE_LENGTH - titleLength} симв.`);
-  } else if (titleLength > MAX_TITLE_LENGTH) {
-    offerTitleInput.setCustomValidity(`Удалите лишние ${titleLength - MAX_TITLE_LENGTH} симв.`);
-  } else {
-    offerTitleInput.setCustomValidity('');
-  }
-  offerTitleInput.reportValidity();
-});
-
-// Проверка соответствия количества комнат количеству гостей.
-const ROOM_GUEST_MAPPER = {
-  1: ['1'],
-  2: ['1', '2'],
-  3: ['1', '2', '3'],
-  100: ['0'],
-};
-
-const roomNumber = document.querySelector('#room_number');
-const guestNumber = document.querySelector('#capacity');
-
-const checkRoomsCapacity = () => {
-  const selectedRooms = roomNumber.value;
-  const selectedGuests = guestNumber.value;
-  if (!ROOM_GUEST_MAPPER[selectedRooms].includes(selectedGuests)) {
-    guestNumber.setCustomValidity(`${selectedRooms}комн. - для ${ROOM_GUEST_MAPPER[selectedRooms].join(' или ')} гостей.`);
-  } else {
-    guestNumber.setCustomValidity('');
-  }
-};
-
-roomNumber.addEventListener('change', checkRoomsCapacity);
-guestNumber.addEventListener('change', checkRoomsCapacity);
-
-const typeInput = document.querySelector('#type');
-const timeIn = document.querySelector('#timein');
-const priceInput = document.querySelector('#price');
-
-const showError = () => {
-  if(typeInput.validity.valueMissing) {
-    typeInput.insertAdjacentHTML('afterend', '<p class="error-message">Заполните поле "Тип жилья".</p>');
-  }
-  if(timeIn.validity.valueMissing) {
-    timeIn.insertAdjacentElement('afterend', '<p class="error-message">Заполните поле "Время заезда и выезда".</p>');
-  }
-  if(priceInput.validity.rangeOverflow) {
-    priceInput.insertAdjacentElement('afterend', '<p class="error-message">Не может быть больше 1 000 000.</p>');
-  }
-};
-
-const form = document.querySelector('.ad-form');
-form.addEventListener('submit', () => {
-  if(!form.validity.valid) {
-    checkRoomsCapacity();
-    showError();
-    Event.preventDefault();
-  }
-});
