+++ js/main.js
@@ -1,2 +1,3 @@
 import './form.js';
+import './form-valid.js';
 import './card.js';
